+++ Makefile
@@ -4,18 +4,26 @@ INC = -I "./"
 
 
 
+
 RayTracing: $(OBJ)
+
 	g++ $(OBJ) -o RayTracing.exe
+
 
 	rm -f $(OBJ)
 
 
 
+
+
 main.o:
 
+	
 	g++ -c main.cpp $(INC)
 
 
 
+
 clean:
+
 	rm -f $(OBJ) RayTracing
