+++ main.cpp
@@ -22,294 +22,12 @@
 #include "Plane.h"
 #include "Triangle.h"
 
+#include "Essentials.h"
+#include "SceneParser.h"
+
 using namespace std;
 
-struct RGBType
-{
-    float r;
-    float g;
-    float b;
-};
-
-void savebmp (const char *filename, int w, int h, int dpi, RGBType *data)
-{
-    FILE *f;
-    int k = w*h;
-    int s = 3*k;
-    int filesize = 54 + s;
-
-    double factor = 39.375;
-    int m = static_cast<int>(factor);
-
-    int ppm = dpi*m;
-
-    unsigned char bmpfileheader[14] = {'B', 'M', 0,0,0,0, 0,0,0,0, 54,0,0,0};
-    unsigned char bmpinfoheader[40] = {40,0,0,0, 0,0,0,0, 0,0,0,0, 1,0,24,0};
-
-    bmpfileheader[ 2] = (unsigned char )(filesize);
-    bmpfileheader[ 3] = (unsigned char )(filesize>>8);
-    bmpfileheader[ 4] = (unsigned char )(filesize>>16);
-    bmpfileheader[ 5] = (unsigned char )(filesize>>24);
-
-    bmpinfoheader[ 4] = (unsigned char)(w);
-    bmpinfoheader[ 5] = (unsigned char)(w>>8);
-    bmpinfoheader[ 6] = (unsigned char)(w>>16);
-    bmpinfoheader[ 7] = (unsigned char)(w>>24);
-
-    bmpinfoheader[ 8] = (unsigned char)(h);
-    bmpinfoheader[ 9] = (unsigned char)(h>>8);
-    bmpinfoheader[10] = (unsigned char)(h>>16);
-    bmpinfoheader[11] = (unsigned char)(h>>24);
-
-    bmpinfoheader[21] = (unsigned char)(s);
-    bmpinfoheader[22] = (unsigned char)(s>>8);
-    bmpinfoheader[23] = (unsigned char)(s>>16);
-    bmpinfoheader[24] = (unsigned char)(s>>24);
-
-    bmpinfoheader[25] = (unsigned char)(ppm);
-    bmpinfoheader[26] = (unsigned char)(ppm>>8);
-    bmpinfoheader[27] = (unsigned char)(ppm>>16);
-    bmpinfoheader[28] = (unsigned char)(ppm>>24);
-
-    bmpinfoheader[29] = (unsigned char)(ppm);
-    bmpinfoheader[30] = (unsigned char)(ppm>>8);
-    bmpinfoheader[31] = (unsigned char)(ppm>>16);
-    bmpinfoheader[32] = (unsigned char)(ppm>>24);
-
-    f=fopen(filename,"wb");
-
-    fwrite(bmpfileheader,1,14,f);
-    fwrite(bmpinfoheader,1,40,f);
-
-    for(int i = 0; i < k; i++)
-    {
-        double red = (data[i].r)*255;
-        double green = (data[i].g)*255;
-        double blue = (data[i].b)*255;
-
-        unsigned char color[3]={(unsigned char)floor(blue),(unsigned char)floor(green),(unsigned char)floor(red)};
-
-        fwrite(color,1,3,f);
-    }
-
-    fclose(f);
-}
-
-int winningObjectIndex(vector<double> object_intersections)
-{
-    // return the index of the winning intersecition
-    int index_of_minimum_value;
-
-    //prevent unnessary calculations
-    if (object_intersections.size() == 0)
-    {
-        //if there are no intersections
-        return -1;
-    }
-    else if(object_intersections.size() == 1)
-    {
-        //
-        if(object_intersections.at(0) > 0)
-        {
-            //if that intersection is greater than zero then its our
-            //index of minimum value
-            return 0;
-        }
-        else
-        {
-            //otherwise the only intersection value is negative
-            return -1;
-        }
-    }
-    else
-    {
-        //otherwise there is more than one intersection
-        //first find the maximum value
-        double max = 0;
-        for (int i = 0; i < object_intersections.size(); i++)
-        {
-            if (max < object_intersections.at(i))
-            {
-                max = object_intersections.at(i);
-            }
-        }
-        // the starting from the maxmimum value to find the minimum positive value
-        if (max > 0)
-        {
-            //we only want positive intersections
-            for (int index = 0; index < object_intersections.size(); index++)
-            {
-                if (object_intersections.at(index) > 0 && object_intersections.at(index) <= max)
-                {
-                    max = object_intersections.at(index);
-                    index_of_minimum_value = index;
-                }
-            }
-            return index_of_minimum_value;
-        }
-        else
-        {
-            //all the intersections were negative
-            return -1;
-        }
-    }
-
-}
-
-Color getColorAt(Vect intersection_position,
-                 Vect intersecting_ray_direction,
-                 vector<Object*> scene_objects,
-                 int index_of_winning_object,
-                 vector<Source*> light_sources,
-                 double accuracy,
-                 double ambientlight)
-{
-    Color winning_object_color = scene_objects.at(index_of_winning_object)->getColor();
-    Vect winning_object_normal = scene_objects.at(index_of_winning_object)->getNormalAt(intersection_position);
-
-    if(winning_object_color.getColorSpecial() == 2)
-    { // checkered/tile floor pattern
-        int square = (int)floor((intersection_position).getVectX())+
-                     (int)floor((intersection_position).getVectZ());
-        if((square%2) == 0)
-        { //black tile
-            winning_object_color.setColorRed(0);
-            winning_object_color.setColorBlue(0);
-            winning_object_color.setColorGreen(0);
-        }
-        else
-        {
-            //white tile
-            winning_object_color.setColorBlue(1);
-            winning_object_color.setColorGreen(1);
-            winning_object_color.setColorRed(1);
-        }
-    }
-
-    Color final_color = winning_object_color.colorScalar(ambientlight);
-
-    if (winning_object_color.getColorSpecial() > 0 && winning_object_color.getColorSpecial() <= 1)
-    { //reflection from objects with specular intensity
-        double dot1 = winning_object_normal.dotProduct(
-                        intersecting_ray_direction.negative());
-        Vect scalar1 = winning_object_normal.vectMult(dot1);
-        Vect add1 = scalar1.vectAdd(intersecting_ray_direction);
-        Vect scalar2 = add1.vectMult(2);
-        Vect add2 = intersecting_ray_direction.negative().vectAdd(scalar2);
-        Vect reflection_direction = add2.normalize();
-
-        Ray reflection_ray (intersection_position,reflection_direction);
-
-        //determine what the ray intersects with first
-        vector<double> reflection_intersections;
-
-        for (int reflection_index = 0; reflection_index < scene_objects.size(); reflection_index++)
-        {
-            reflection_intersections.push_back(scene_objects.at(reflection_index)->
-                                               findIntersection(reflection_ray));
-        }
-
-        int index_of_winning_object_with_reflection = winningObjectIndex(reflection_intersections);
-
-        if (index_of_winning_object_with_reflection != -1)
-        {   //reflection ray missed everything else
-            if (reflection_intersections.at(index_of_winning_object_with_reflection)>accuracy)
-            { // determine the position and direction at the point of intersection with
-                //the reflection ray. the ray only affects the color if it reflected off something
-                Vect reflection_intersection_position = intersection_position.vectAdd(
-                                                        reflection_direction.vectMult(
-                                                        reflection_intersections.at(
-                                                        index_of_winning_object_with_reflection)));
-                Vect reflection_intersection_ray_direction = reflection_direction;
-
-                Color reflection_intersection_color = getColorAt(reflection_intersection_position,
-                                                                 reflection_intersection_ray_direction,
-                                                                 scene_objects,
-                                                                 index_of_winning_object_with_reflection,
-                                                                 light_sources,
-                                                                 accuracy, ambientlight);
-                final_color = final_color.colorAdd(reflection_intersection_color.colorScalar(
-                                                    winning_object_color.getColorSpecial()));
-
-
-            }
-        }
-    }
-
-    for (int light_index = 0; light_index < light_sources.size(); light_index++)
-    {
-        Vect light_direction = light_sources.at(light_index)->getLightPosition().vectAdd(
-                                                     intersection_position.negative()).normalize();
-
-        float cosine_angle = winning_object_normal.dotProduct(light_direction);
-
-        if(cosine_angle > 0)
-        {
-            //test for shadows
-            bool shadowed = false;
-
-            Vect distance_to_light = light_sources.at(light_index)->getLightPosition().vectAdd(
-                                                    intersection_position.negative());
-            float distance_to_light_magnitude = distance_to_light.magnitude();
-
-            Ray shadow_ray(intersection_position,
-                           light_sources.at(light_index)->getLightPosition().vectAdd(
-                                                  intersection_position.negative()).normalize());
-            vector<double> secondary_intersections;
-
-            for (int object_index = 0; object_index < scene_objects.size() && shadowed == false; object_index++)
-            {
-                secondary_intersections.push_back(scene_objects.at(object_index)->
-                                                  findIntersection(shadow_ray));
-            }
-
-            for(int c = 0; c < secondary_intersections.size(); c++)
-            {
-                if (secondary_intersections.at(c) > accuracy)
-                {
-                    if (secondary_intersections.at(c) <= distance_to_light_magnitude)
-                    {
-                        shadowed = true;
-                    }
-                    break;
-                }
-            }
-
-            if (shadowed == false)
-            {
-                final_color = final_color.colorAdd(winning_object_color.colorMult(
-                                                   light_sources.at(light_index)->getColor()).
-                                                   colorScalar(cosine_angle));
-                if(winning_object_color.getColorSpecial() > 0 && winning_object_color.getColorSpecial() <= 1)
-                { //entre 0 e 1 é shininess, outros valores são outras coisas
-                    double dot1 = winning_object_normal.dotProduct(
-                                  intersecting_ray_direction.negative());
-                    Vect scalar1 = winning_object_normal.vectMult(dot1);
-                    Vect add1 = scalar1.vectAdd(intersecting_ray_direction);
-                    Vect scalar2 = add1.vectMult(2);
-                    Vect add2 = intersecting_ray_direction.negative().vectAdd(scalar2);
-                    Vect reflection_direction = add2.normalize();
-
-                    double specular = reflection_direction.dotProduct(light_direction);
-                    if (specular > 0)
-                    {
-                        specular = pow(specular, 10);
-                        final_color = final_color.colorAdd(
-                                      light_sources.at(light_index)->
-                                      getColor().colorScalar(
-                                      specular*winning_object_color.getColorSpecial()));
-                    }
-                }
-            }
-        }
-    }
-
-    return final_color.clip();
-}
-
-vector<Object*> scene_objects;
-
-void makeCube (Vect corner1, Vect corner2, Color color)
+void makeCube (Vect corner1, Vect corner2, Color color, vector<Object*>& scene_objects)
 {
         //corner1
         double c1x = corner1.getVectX();
@@ -350,24 +68,32 @@ void makeCube (Vect corner1, Vect corner2, Color color)
 }
 
 
-int main(int argc, char *argv[])
+int main()
 {
     cout << "rendering..." << endl;
 
     clock_t t1, t2;
+    time_t now;
     t1 = clock();
 
+    vector<Object*> scene_objects;
+
+    char arquivo[25];
+    string path = "T:\\GitHub\\RayTracing\\Resultados\\";
+    string filename2;
+
     int dpi=72;
-    int width=640;
-    int height=480;
+    int width, height, aadepth;
+    double ambientlight;
+    Vect campos, look_at;// (3, 1.5, -4);
+    //Vect look_at (0,0,0);
+    SceneParser(height, width, aadepth, ambientlight, campos, look_at, scene_objects);
     int n = width*height;
     int thisone;
     RGBType *pixels = new RGBType[n];
 
-    int aadepth = 1;
-    double aathreshold = 0.1;
+   // double aathreshold = 0.1;
     double aspectratio = (double)width/(double)height;
-    double ambientlight = 0.2;
     double accuracy = 0.000001;
 
     Vect O (0,0,0);
@@ -377,9 +103,7 @@ int main(int argc, char *argv[])
 
     Vect location2 (1.75, 0, 0);
 
-    Vect campos (3, 1.5, -4);
 
-    Vect look_at (0,0,0);
     Vect diff_btw (campos.getVectX() - look_at.getVectX(),
                    campos.getVectY() - look_at.getVectY(),
                    campos.getVectZ() - look_at.getVectZ());
@@ -406,16 +130,16 @@ int main(int argc, char *argv[])
     //scene objects
     //Sphere scene_sphere (O, 1, pretty_green);
     Plane scene_plane(Y, -1, checkered); //Y É A NORMAL????
-    Sphere scene_sphere2 (location2, 0.5, maroon);
+    //Sphere scene_sphere2 (location2, 0.5, maroon);
     //Triangle scene_triangle(Vect (3, 0, 0), Vect (0, 3, 0), Vect (0, 0, 3), orange);
 
 
     //scene_objects.push_back(dynamic_cast<Object*>(&scene_sphere));
     scene_objects.push_back(dynamic_cast<Object*>(&scene_plane));
-    scene_objects.push_back(dynamic_cast<Object*>(&scene_sphere2));
+    //scene_objects.push_back(dynamic_cast<Object*>(&scene_sphere2));
     //scene_objects.push_back(dynamic_cast<Object*>(&scene_triangle));
 
-    makeCube(Vect (1,1,1), Vect(-1,-1,-1), orange);
+    //makeCube(Vect (1,1,1), Vect(-1,-1,-1), orange, scene_objects);
 
     double xamnt, yamnt;
     //double tempRed, tempGreen, tempBlue;
@@ -497,7 +221,7 @@ int main(int argc, char *argv[])
 
                     vector<double> intersections;
 
-                    for (int index = 0; index < scene_objects.size(); index++)
+                    for (unsigned index = 0; index < scene_objects.size(); index++)
                     {
                         intersections.push_back(scene_objects.at(index)->findIntersection(cam_ray));
                     }
@@ -567,8 +291,15 @@ int main(int argc, char *argv[])
             pixels[thisone].b=avgBlue;
         }
     }
-    string filename = "teste.bmp";
-    savebmp(filename.c_str(), width, height, dpi, pixels);
+
+    now = time(NULL);
+    strftime(arquivo,25,"%Y-%m-%d %H.%M.%S.bmp",localtime(&now));
+
+    filename2 = arquivo;
+    filename2 = path+filename2;
+    //cout << filename2 << endl;
+
+    savebmp(filename2.c_str(), width, height, dpi, pixels);
 
     delete pixels;//, tempRed, tempGreen, tempBlue;
 
@@ -576,6 +307,6 @@ int main(int argc, char *argv[])
     float diff = ((float)t2 - (float)t1)/1000;
 
     cout << diff << " seconds" << endl;
-    cout << "Output: " << filename << endl;
+    cout << "Output: " << filename2.c_str() << endl;
     return 0;
 }
