+++ scenefile.txt
@@ -1,0 +1,18 @@
+//DO NOT MODIFY ANY COMMENTED LINE! (LINES THAT BEGIN WITH '//')
+//Width of the image
+1600
+//Height of the image
+900
+//AA depth of the image, recommended maximum 4
+1
+//Ambient light of the image
+0.2
+//Camera position in a X Y Z format, please
+3 1.5 -4
+//Where do you want the camera to look at, in X Y Z format
+0 0 0
+//How many spheres do you want?
+2
+//Sphere positions in X Y Z R format, where R is radius and color in R G B S format
+1.75 0 0 0.5 0.5 0.25 0.25 0
+0 0 0 1 0.5 1 0.5 0.3
